+++ src/atomiclines/__init__.py
@@ -2,9 +2,15 @@
 
 from atomiclines.atomiclinereader import AtomicLineReader
 from atomiclines.backgroundtask import BackgroundTask
-from atomiclines.lineprocessor import LineProcessor
+from atomiclines.lineprocessor import LineHolder, LineProcessor
 from atomiclines.log import logger, try_load_config_from_environment
 
-__all__ = ["AtomicLineReader", "BackgroundTask", "LineProcessor", "logger"]
+__all__ = [
+    "AtomicLineReader",
+    "BackgroundTask",
+    "LineHolder",
+    "LineProcessor",
+    "logger",
+]
 
 try_load_config_from_environment("ATOMICLINES_LOG_CONFIG")